config: extract univeral_read() into a helper
The plan is to re-use it in a next commit.

Part of #9506

NO_DOC=no public API changes
NO_CHANGELOG=see NO_DOC
NO_TEST=see NO_DOC
