+++ src/box/CMakeLists.txt
@@ -58,6 +58,7 @@ lua_source(lua_sources lua/config/instance_config.lua     config_instance_config
 lua_source(lua_sources lua/config/source/env.lua          config_source_env_lua)
 lua_source(lua_sources lua/config/source/file.lua         config_source_file_lua)
 lua_source(lua_sources lua/config/utils/expression.lua    config_utils_expression_lua)
+lua_source(lua_sources lua/config/utils/file.lua          config_utils_file_lua)
 lua_source(lua_sources lua/config/utils/log.lua           config_utils_log_lua)
 lua_source(lua_sources lua/config/utils/schema.lua        config_utils_schema_lua)
 lua_source(lua_sources lua/config/utils/snapshot.lua      config_utils_snapshot_lua)
